requirements.txt
Django==2.1.7
gunicorn==19.9.0
mysqlclient==1.4.2.post1
pytz==2018.9
PyYAML==3.13
six==1.12.0
SQLAlchemy==1.3.0
SQLAlchemy-Utils==0.33.11
whitenoise==4.1.2
